Make "new topics count" async
We had a "there are X new topics" notification already, but it was only
refreshed during page load - so unless a user started scrolling, never.

We also calculated this for every request, which could be constly for
some search queries, greatly increasing page load time.

This commit extracts it to work asynchronously, polling regularly, which
decouples it from the main page load time and also makes it more useful.