Fix topics caching bug
We used topics.last in the watermark which is the last topic on the page
instead of the first which is the newest topic on the page.

This caused the page not to properly refresh if the last topic changed,
but only by a thread already on the page receiveing a new message.