Fixing unread message counting bug
We assumed that with message ranges (id1,id2) we can count the number of
messages in the range by id2-id1. This is of course wrong because some
of the ids in the range can belong to other threads, causing the code to
significantly overcount read messages.

To fix this problem this commit introduces a new column to the ranges,
so that we store ranges in the form of (id1,id2,count), and can count
the read message count reliable for threads by a simple group by
thread_id sum(count).